# backend/api/build/resources/main/openapi/dopashift-api-v1.yml
openapi: 3.0.3
info:
  title: DopaShift API
  description: |
    REST API for the DopaShift multi-goal life-tracking platform.
    
    ## Versioning Policy
    - All endpoints are prefixed with `/v1/`
    - Previous API versions are supported for 6 months after a new version is released.
    - Breaking changes will only be introduced in new major versions.
    
    ## Authentication
    All endpoints (except /v1/auth/*) require a valid Bearer token in the Authorization header.
    Tokens are obtained via the Keycloak-proxied authentication endpoints.
  version: 1.0.0
  contact:
    name: DopaShift Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.dopashift.app
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/reminders:
    post:
      tags: [Reminders]
      operationId: createReminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
    get:
      tags: [Reminders]
      operationId: listReminders
      responses:
        '200':
          description: All reminders for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderResponse'
  /v1/reminders/{id}:
    put:
      tags: [Reminders]
      operationId: updateReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          description: Reminder updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
    delete:
      tags: [Reminders]
      operationId: deleteReminder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Reminder deleted

  # === Interception Rules ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateReminderRequest:
      type: object
      required: [entityType, entityId, scheduledTime]
      properties:
        entityType:
          type: string
          enum: [DAILY_TODO, GOAL_CHECKLIST, HABIT_CHECKPOINT]
        entityId:
          type: string
          format: uuid
        scheduledTime:
          type: string
          example: "09:00"
        scheduledDate:
          type: string
          format: date
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
        conditionType:
          type: string
          enum: [ENTITY_INCOMPLETE]
        escalationIntervalMinutes:
          type: integer
          minimum: 1
        maxEscalations:
          type: integer
          minimum: 0
    UpdateReminderRequest:
      type: object
      properties:
        scheduledTime:
          type: string
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
        isActive:
          type: boolean
    RecurrenceRule:
      type: object
      properties:
        type:
          type: string
          enum: [DAILY, SPECIFIC_WEEKDAYS, WEEKLY, MONTHLY, CUSTOM_INTERVAL]
        weekdays:
          type: array
          items:
            type: string
            enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
        intervalDays:
          type: integer
    ReminderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        scheduledTime:
          type: string
        recurrence:
          $ref: '#/components/schemas/RecurrenceRule'
        isActive:
          type: boolean
        escalationIntervalMinutes:
          type: integer
        maxEscalations:
          type: integer

    # === Interception ===
...